Fix for bug 9157 (trace back window appears when setting up accounts)
git-svn-id: http://svn.osafoundation.org/chandler/trunk@14327 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/application/dialogs/AccountPreferences.py
@@ -799,7 +799,8 @@ class AccountPreferencesDialog(wx.Dialog):
                         folders.remove(folder)
                         folder.delete()
 
-                item.password.delete()
+                if hasattr(item, 'password'):
+                    item.password.delete()
                 item.delete(recursive=True)
 
     def __ApplyCancellations(self):
